Move CC Problem report fixture into `internal-integ-testing`
