+++ testing/smoke-ide-test/src/smokeIdeTest/groovy/org/gradle/ide/sync/IsolatedProjectsGradleceptionSyncTest.groovy
@@ -18,7 +18,7 @@ package org.gradle.ide.sync
 
 
 import org.gradle.ide.starter.IdeScenarioBuilder
-import org.gradle.ide.sync.fixtures.IsolatedProjectsIdeSyncFixture
+import org.gradle.integtests.fixtures.configurationcache.ConfigurationCacheProblemReportFixture
 import org.gradle.integtests.fixtures.AvailableJavaHomes
 import org.gradle.test.fixtures.Flaky
 import org.gradle.test.fixtures.file.TestFile
@@ -27,7 +27,7 @@ import org.gradle.test.fixtures.file.TestFile
 class IsolatedProjectsGradleceptionSyncTest extends AbstractIdeSyncTest {
 
     private TestFile gradleCheckout = testDirectory.createDir("gradle-checkout")
-    private IsolatedProjectsIdeSyncFixture fixture = new IsolatedProjectsIdeSyncFixture(gradleCheckout)
+    private ConfigurationCacheProblemReportFixture fixture = new ConfigurationCacheProblemReportFixture(gradleCheckout)
 
     def "can sync gradle/gradle build without problems"() {
         given:
